Rename file style css
