+++ src/components/Header/Header.tsx
@@ -1,4 +1,4 @@
-import { HeaderSection, HeaderContainer } from "./HeaderStyles";
+import { HeaderSection, HeaderContainer } from "./Header.style";
 import  Navbar from "./Navbar";
 
 const Header = () => {
